(hack) lots of gross casts to get it compiling
There's a few unit conversion problems that have cropped up, chiefly:

* Pixel is defined as i32
  * Represents a point on the screen, can be negative for fun and profit
  * Also can be used to represent a length or size of an object

SDL is now using i32 for points on screen, and u32 for lengths.
I will either have to introduce a new unit type to handle this, or live with the casts.

(There's an additional problem in animated sprites; it's using number of frames (u64) to determine a length (u32)
 that becomes an offset (i32)... which means animations are limited to 4 billion frames.)

+++ src/backdrop.rs
@@ -27,14 +27,14 @@ impl FixedBackdrop {
 
 		while units::Pixel(x) < game::SCREEN_WIDTH.to_pixel() {
 			while units::Pixel(y) < game::SCREEN_HEIGHT.to_pixel() {
-				let src  = Rect::new(0, 0, tile_size, tile_size);
-				let dest = Rect::new(x, y, tile_size, tile_size);
+				let src  = Rect::new(0, 0, tile_size as u32, tile_size as u32);
+				let dest = Rect::new(x, y, tile_size as u32, tile_size as u32);
 
 				graphics.blit_surface(&self.surface_id[..], &src, &dest);
-				y+= tile_size;
+				y+= tile_size as i32;
 			}
 
-			x += tile_size;
+			x += tile_size as i32;
 			y = 0;
 		}
 	}
